+++ src/logic/ConfigManager.ts
