+++ src/logic/IPseudonymResult.ts
